requirements/requirements.txt
discord.py
paramiko
pyyaml
